Update 2017-06-05-free-open-source-api-doc-tools.md
We usually call ourselves "developer portal creation specialists", instead of "experts", so I edited this just to be consistent. (If you think "experts" sounds better, please feel free to change it back.)
I agree with putting the definitions and who we wrote this post for at the end, because your audience probably already has this background knowledge.